+++ tests/cases/properties/class-papi-property-editor-test.php
@@ -15,13 +15,32 @@ class Papi_Property_Editor_Test extends Papi_Property_Test_Case {
 		return "<p>a bit of text with html tags hello, world</p>\n";
 	}
 
+	public function test_mce_buttons() {
+		$property = papi_property( [
+			'title' => 'Editor',
+			'type'  => 'editor'
+		] );
+		$this->assertEmpty( $property->mce_buttons() );
+
+		global $wp_current_filter;
+		$wp_current_filter[] = 'mce_buttons';
+		$property = papi_property( [
+			'title'    => 'Editor',
+			'type'     => 'editor',
+			'settings' => [
+				'mce_buttons' => ['bold']
+			]
+		] );
+		$this->assertSame( ['bold'], $property->mce_buttons() );
+	}
+
 	public function test_property_default_value() {
-		$post_id = $this->factory->post->create(['post_status' => 'auto-draft']);
+		$post_id = $this->factory->post->create( ['post_status' => 'auto-draft'] );
 		$property = clone $this->property;
 		$property->set_post_id( $post_id );
-		$property->set_option('default', 'You need to update your profile, <a href="/profile/update">click here!</a>');
+		$property->set_option( 'default', 'You need to update your profile, <a href="/profile/update">click here!</a>' );
 		papi_render_property( $property );
-		$this->expectOutputRegex('/\>You need to update your profile, \<a href\=\"\/profile\/update\"\>click here!\<\/a\>\<\/textarea\>/');
+		$this->expectOutputRegex( '/\>You need to update your profile, \<a href\=\"\/profile\/update\"\>click here!\<\/a\>\<\/textarea\>/' );
 	}
 
 	public function test_property_format_value() {